Moves the structures out of the main file of the part one solution of the day 18
